add image selection to notification creation page
The new button is named image_selector_button which has event added as action on mouse click. The event opens the filechooser with the current stage as parent . The file is selected and the variable imageisselected is set to true and file_name is also set . The firebase storage client is used in line 121 which returns the bucket of firebase storage . Create method is called in bucket which creates a file at the root of the the firebase storage with the same name and with access control set as public read and returns the blob to file. getMedialink() method is called on blob object which returns the public url of the blob. The public url is added to firebase message with key as "img".